requirements.txt
customtkinter
pywin32
keyboard
screeninfo
Pillow
numpy
watchdog
